Создала МОКи для офферов и комментариев. Статично, по 4 шт.
