+++ src/mocks/offers.ts
@@ -1,0 +1,164 @@
+const offer1 =
+  {
+    'bedrooms': 3,
+    'city': {
+      'location': {
+        'latitude': 52.370216,
+        'longitude': 4.895168,
+        'zoom': 10
+      },
+      'name': 'Amsterdam'
+    },
+    'description': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
+    'goods': [
+      'Heating'
+    ],
+    'host': {
+      'avatarUrl': 'img/1.png',
+      'id': 3,
+      'isPro': true,
+      'name': 'Angelina'
+    },
+    'id': 1,
+    'images': [
+      'img/1.png'
+    ],
+    'isFavorite': true,
+    'isPremium': true,
+    'location': {
+      'latitude': 52.35514938496378,
+      'longitude': 4.673877537499948,
+      'zoom': 8
+    },
+    'maxAdults': 4,
+    'previewImage': 'img/1.png',
+    'price': 320,
+    'rating': 4.8,
+    'title': 'Beautiful & luxurious apartment at great location',
+    'type': 'apartment'
+  };
+
+const offer2 = {
+  'bedrooms': 2,
+  'city': {
+    'location': {
+      'latitude': 52.370216,
+      'longitude': 4.895168,
+      'zoom': 8
+    },
+    'name': 'Amsterdam'
+  },
+  'description': 'Nice studio in Amsterdam.',
+  'goods': [
+    'Heating'
+  ],
+  'host': {
+    'avatarUrl': 'img/2.png',
+    'id': 5,
+    'isPro': true,
+    'name': 'Marusya'
+  },
+  'id': 2,
+  'images': [
+    'img/2.png'
+  ],
+  'isFavorite': false,
+  'isPremium': false,
+  'location': {
+    'latitude': 52.35514938496378,
+    'longitude': 4.673877537499948,
+    'zoom': 8
+  },
+  'maxAdults': 2,
+  'previewImage': 'img/1.png',
+  'price': 320,
+  'rating': 4.5,
+  'title': 'Nice studio at great location',
+  'type': 'studio'
+};
+
+const offer3 = {
+  'bedrooms': 1,
+  'city': {
+    'location': {
+      'latitude': 52.370216,
+      'longitude': 4.895168,
+      'zoom': 8
+    },
+    'name': 'Amsterdam'
+  },
+  'description': 'Nice studio in Amsterdam.',
+  'goods': [
+    'Heating'
+  ],
+  'host': {
+    'avatarUrl': 'img/1.png',
+    'id': 5,
+    'isPro': true,
+    'name': 'Marusya'
+  },
+  'id': 3,
+  'images': [
+    'img/3.png'
+  ],
+  'isFavorite': false,
+  'isPremium': false,
+  'location': {
+    'latitude': 52.35514938496378,
+    'longitude': 4.673877537499948,
+    'zoom': 8
+  },
+  'maxAdults': 1,
+  'previewImage': 'img/1.png',
+  'price': 150,
+  'rating': 4.5,
+  'title': 'Nice studio at great location',
+  'type': 'studio'
+};
+
+const offer4 = {
+  'bedrooms': 2,
+  'city': {
+    'location': {
+      'latitude': 52.370216,
+      'longitude': 4.895168,
+      'zoom': 8
+    },
+    'name': 'Amsterdam'
+  },
+  'description': 'Nice studio in Amsterdam.',
+  'goods': [
+    'Heating'
+  ],
+  'host': {
+    'avatarUrl': 'img/1.png',
+    'id': 5,
+    'isPro': true,
+    'name': 'Marusya'
+  },
+  'id': 4,
+  'images': [
+    'img/3.png'
+  ],
+  'isFavorite': false,
+  'isPremium': false,
+  'location': {
+    'latitude': 52.35514938496378,
+    'longitude': 4.673877537499948,
+    'zoom': 8
+  },
+  'maxAdults': 1,
+  'previewImage': 'img/1.png',
+  'price': 250,
+  'rating': 4.5,
+  'title': 'Nice studio at great location',
+  'type': 'studio'
+};
+
+const OFFERS = [
+  offer1, offer2, offer3, offer4
+];
+
+// const OFFERS = [];
+
+export {OFFERS};
