+++ src/mocks/reviews.ts
@@ -1,0 +1,58 @@
+const review1 = {
+  'comment': 'A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
+  'date': 'Mon Oct 30 2023 15:27:12 GMT+0500 (Екатеринбург, стандартное время)',
+  'id': 1,
+  'rating': 4,
+  'user': {
+    'avatarUrl': 'img/1.png',
+    'id': 1,
+    'isPro': false,
+    'name': 'Ali.Baba'
+  }
+};
+
+const review2 = {
+  'comment': 'comment N2. A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
+  'date': 'Mon Oct 30 2023 15:27:12 GMT+0500 (Екатеринбург, стандартное время)',
+  'id': 2,
+  'rating': 3,
+  'user': {
+    'avatarUrl': 'img/1.png',
+    'id': 2,
+    'isPro': false,
+    'name': 'Oliver.conner'
+  }
+};
+
+const review3 = {
+  'comment': 'comment N3. A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
+  'date': 'Mon Oct 30 2023 15:27:12 GMT+0500 (Екатеринбург, стандартное время)',
+  'id': 3,
+  'rating': 5,
+  'user': {
+    'avatarUrl': 'img/1.png',
+    'id': 3,
+    'isPro': false,
+    'name': 'Marusya.ivanova'
+  }
+};
+
+const review4 = {
+  'comment': 'comment N4. A quiet cozy and picturesque that hides behind a a river by the unique lightness of Amsterdam.',
+  'date': 'Mon Oct 30 2023 15:27:12 GMT+0500 (Екатеринбург, стандартное время)',
+  'id': 4,
+  'rating': 3,
+  'user': {
+    'avatarUrl': 'img/1.png',
+    'id': 2,
+    'isPro': false,
+    'name': 'Oliver.conner'
+  }
+};
+
+const REVIEWS = [
+  review1, review2, review3, review4
+];
+
+export {REVIEWS};
+
